niffler_tests_python/requirements.txt
python-dotenv
pytest
requests
Faker
python-dateutil
selenium
webdriver-manager
pydantic
allure-pytest
requests-toolbelt
psycopg2-binary
sqlalchemy
sqlmodel